requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
requests==2.31.0
psycopg2-binary==2.9.10
apscheduler==3.10.1
charset-normalizer==3.2.0
ftfy==6.1.1